analyzer_service/requirements.txt
python-dotenv==0.15.0
fastapi==0.103.2